Fix unable to invoke `pinentry-gnome3` from git
invoke pinentry-gnome3 because it cannot
refer to the DISPLAY environment variable.
This occurs because systemd 226 and later
versions run dbus as a user instance, which
does not inherit environment variables set
for the desktop environment.

The `dbus-update-activation-environment
--systemd` command allows setting
environment variables for systemd user
services and dbus sessions. This commit
adds this command to lightdm's
sessionCommands to address the issue.